requirements.txt
fastapi
uvicorn
pandas
numpy
psycopg2-binary
scikit-learn
joblib
requests
cachetools
python-dotenv
upstox-python
pyotp
selenium
webdriver-manager
pydantic
yfinance
sqlalchemy
xgboost
